move log package to internal
