+++ tools/ghx/runner/cmd_executor.go
@@ -12,8 +12,8 @@ import (
 	"strings"
 
 	"github.com/aweris/gale/internal/core"
+	"github.com/aweris/gale/internal/log"
 	"github.com/aweris/gale/tools/ghx/actions"
-	"github.com/aweris/gale/tools/ghx/log"
 )
 
 type CmdExecutor struct {
