+++ tools/ghx/runner/runner.go
@@ -5,8 +5,8 @@ import (
 	"fmt"
 
 	"github.com/aweris/gale/internal/core"
+	"github.com/aweris/gale/internal/log"
 	"github.com/aweris/gale/tools/ghx/actions"
-	"github.com/aweris/gale/tools/ghx/log"
 )
 
 type Runner struct {
