+++ tools/ghx/runner/runner_step.go
@@ -8,8 +8,8 @@ import (
 	"github.com/aweris/gale/internal/config"
 	"github.com/aweris/gale/internal/core"
 	"github.com/aweris/gale/internal/fs"
+	"github.com/aweris/gale/internal/log"
 	"github.com/aweris/gale/tools/ghx/actions"
-	"github.com/aweris/gale/tools/ghx/log"
 )
 
 type Step interface {
